+++ CHANGELOG.md
@@ -423,6 +423,14 @@ This update overhauls Predator's dashcam functionality and adds various features
 - Added headless operation mode, where all user input prompts are skipped.
 
 
+## Version 10.0.1
+
+March 6th, 2024
+
+- Fixed a bug where the audio file name wouldn't be properly updated when a new dash-cam video segment starts.
+- Warnings now wait for user input before continuing, just like errors.
+
+
 ## Version 10.1
 
 *Release date to be determined*
@@ -435,4 +443,3 @@ This update overhauls Predator's dashcam functionality and adds various features
     - Predator now checks to see if each value in the configuration matches the expected data type.
     - Configuration values that exist in the configuration template file but not in the actual configuration file will show errors.
 - Added performance monitoring to state interface file to allow external programs to see basic performance diagnostics.
-- Fixed a bug where the audio file name wouldn't be properly updated when a new dash-cam video segment starts.